requirements.txt
appdirs==1.4.0
github3.py==0.9.6
packaging==16.8
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
uritemplate==3.0.0
uritemplate.py==3.0.2
